main: Split off events loop handling into a separate file
To make it easier to use as an example.
